+++ R/data.R
@@ -1,0 +1,110 @@
+
+#' CRAN riskmetric Scores on 2023-06-21
+#'
+#' The initial data.frame is generated as a placeholder until a more streamlined
+#' process is implemented. It contains the initial run where all of CRAN was
+#' scored using riskmetric v0.2.1 on 2023-06-21.
+#'
+#' @format A `data.frame` containing 19,715 observations
+#'   (one row per package) and 23 variables, 18 of which are riskmetric
+#'   assessments.
+#'
+#' \describe{
+#'   \item{package}{The name of a R package hosted on CRAN}
+#'   \item{version}{The package's version number}
+#'   \item{riskmetric_run_date}{The date riskmetric was run}
+#'   \item{riskmetric_version}{The version of riskmetric used to derive scores}
+#'   \item{pkg_score}{The package's riskmetric score}
+#' }
+#' All other variables correspond to the riskmetric assessments that exist at
+#'   the time of execution. Please refer to riskmetric's documentation for more
+#'   info.
+#'
+#' @return A data.frame
+"cran_scored_20230621"
+
+
+#' CRAN riskmetric Scores on 2025-08-12
+#'
+#' A data.frame containing all of CRAN was scored using riskmetric v0.2.5 on
+#' 2025-08-12 & R v4.5.1.
+#'
+#' @format A `data.frame`
+#'
+#' \describe{
+#'   \item{package}{The name of a R package hosted on CRAN}
+#'   \item{version}{The package's version number}
+#'   \item{riskmetric_run_date}{The date riskmetric was run}
+#'   \item{riskmetric_version}{The version of riskmetric used to derive scores}
+#'   \item{pkg_score}{The package's riskmetric score}
+#' }
+#'   All other variables correspond to the riskmetric scores that exist at
+#'   the time of execution. Please refer to riskmetric's documentation for more
+#'   info.
+#'
+#' @return A data.frame
+"cran_scored_20250812"
+
+
+#' CRAN riskmetric Assessments on 2025-08-12
+#'
+#' A data.frame containing all of CRAN was assessed using riskmetric v0.2.5 on
+#' 2025-08-12 & R v4.5.1.
+#'
+#' @format A `data.frame`
+#'
+#' \describe{
+#'   \item{package}{The name of a R package hosted on CRAN}
+#'   \item{version}{The package's version number}
+#'   \item{riskmetric_run_date}{The date riskmetric was run}
+#'   \item{riskmetric_version}{The version of riskmetric used to derive scores}
+#' }
+#' All other variables correspond to the riskmetric assessments that exist at
+#'   the time of execution. Please refer to riskmetric's documentation for more
+#'   info.
+#'
+#' @return A data.frame
+"cran_assessed_20250812"
+
+
+#' Most recent riskmetric Scores for CRAN
+#'
+#' A data.frame containing all of CRAN was scored using riskmetric v0.2.5 on
+#' 2025-08-12 & R v4.5.1.
+#'
+#' @format A `data.frame`
+#'
+#' \describe{
+#'   \item{package}{The name of a R package hosted on CRAN}
+#'   \item{version}{The package's version number}
+#'   \item{riskmetric_run_date}{The date riskmetric was run}
+#'   \item{riskmetric_version}{The version of riskmetric used to derive scores}
+#'   \item{pkg_score}{The package's riskmetric score}
+#' }
+#'   All other variables correspond to the riskmetric scores that exist at
+#'   the time of execution. Please refer to riskmetric's documentation for more
+#'   info.
+#'
+#' @return A data.frame
+"cran_scored_latest"
+
+
+#' Most recent riskmetric Assessments for CRAN
+#'
+#' A data.frame containing all of CRAN was assessed using riskmetric v0.2.5 on
+#' 2025-08-12 & R v4.5.1.
+#'
+#' @format A `data.frame`
+#'
+#' \describe{
+#'   \item{package}{The name of a R package hosted on CRAN}
+#'   \item{version}{The package's version number}
+#'   \item{riskmetric_run_date}{The date riskmetric was run}
+#'   \item{riskmetric_version}{The version of riskmetric used to derive scores}
+#' }
+#' All other variables correspond to the riskmetric assessments that exist at
+#'   the time of execution. Please refer to riskmetric's documentation for more
+#'   info.
+#'
+#' @return A data.frame
+"cran_assessed_latest"