+++ NEWS.md
@@ -1,3 +1,8 @@
+# riskscore 0.0.1
+
+* Revamped process to store assessments (in addition to scores) in a tibble.
+
 # riskscore 0.0.0.9000
 
-* set up pkg skeleton with initial `riskscore` placeholder
+* set up pkg skeleton with initial `riskscore` placeholder on 2023-06-21
+